strip out all the AAT weather functions from hpocs and add to script
