+++ src/huntsman/pocs/utils/safety.py
@@ -8,31 +8,6 @@ from astropy.coordinates import get_sun, AltAz
 
 from panoptes.utils.utils import get_quantity_value
 
-import requests
-
-AAT_URL = 'http://aat-ops.anu.edu.au/met/metdata.dat'
-AAT_COLUMNS = ['time',
-               'outside_temp',
-               'inside_temp',
-               'mirror_temp',
-               'outside_dewpoint',
-               'outside_humidity',
-               'pressure',
-               'wind_speed_avg',
-               'wind_gust_max',
-               'wind_direction_avg',
-               'dome_state',
-               'is_raining',
-               'inside_dewpoint',
-               'sky_ambient_diff_C',
-               'sky_ambient_diff_error',
-               'daytime_brightness',
-               'rain_detections_past_10minutes',
-               'wetness_detections_past_10minutes',
-               'rain_since_9am',
-               'sqm_brightness',
-               ]
-
 
 def get_solar_altaz(time, location):
     """ Return the altaz of the Sun at a given time and location.
@@ -98,25 +73,3 @@ def check_solar_separation_safety(observation, location, min_separation, time=No
     separations = get_solar_separation(coord, times, location)
 
     return all([c > min_separation for c in separations])
-
-
-def get_aat_weather(aat_url=AAT_URL, response_columns=AAT_COLUMNS):
-    """Fetch met weather data from AAO weather station.
-
-    Args:
-        aat_url (string, optional): URL to query for weather reading. Defaults to AAT_URL.
-        response_columns (list, optional): List of column names that map onto the values contained
-        in a succesful reponse. Defaults to AAT_COLUMNS.
-    """
-    response = requests.get(aat_url)
-    # raise an exception if response was not successful
-    response.raise_for_status()
-
-    if response.ok:
-        date, raw_data, _ = response.content.decode().split('\n')
-        data = {name: value for name, value in zip(response_columns, raw_data.split('\t'))}
-        data['date'] = date
-    else:
-        data = None
-
-    return data
